version: Set the cluster operator version based on the payload
The new contract with operators is that we report the version when
we reach "level" (have rolled out the version we got handed in the
payload) and are available, which ensures that the CVO can wait
until it sees us completely roll out the payload. This makes upgrade
status much clearer.