Degrade the cluster when an unsupported CSI driver is already installed
When opt-in CSI driver should be installed and the driver is already
installed, but it's not the OpenShift one, Degrade the whole cluster and
wait until user removes the other (community) CSI driver.

This affects only CSI drivers that are installed after TechPreviewNoUpgrade
FeatureSet is set.

We distinguish "our" CSI driver by presence of annotation
"csi.openshift.io/managed".