Don't set Available condition when deploying CSI operator
Currently, we deploy only one replica of the CSI operator Deployment.
This contributes to the situation where to CSO goes unavailable when
draining the node where a CSI operator is running on.

A reasonable way to fix this issue would be to run multiple replicas
of the CSI operator, however, this patch removes the
_CSIDriverOperatorDeploymentAvailable_ condition instead. This condition
was was being set to _False_ when draining the node where the CSI operator
was running on, cascading the status up to the top _Available_ condition
(through the unioning done in the CR controller in CSO).

From now on, the unioning that determines the top _Available_ state relies
on the other _Available_ conditions reported by other controllers and by
the CSI operator.

In addition to that, I introduced a health check for the CSI operator
Deployment to make sure we go _Degraded_ if for some reason the CSI
operator is not deployed successfully.