Add UID to CSO Pod to be able to run with custom SCCs
When there is a custom SCC with "runAsUser: type: runAsNonRoot", CSO pod
can match it. Then the Pod must have a non-root UID assigned to be able to
run.

This moves the CSO Pods from "restricted-v2" SCC to "nonroot-v2" SCC in the
default OCP installation.

This is a workaround for CSO using cluster-admin role, which can allow CSO
to use any SCC. If CSO had specific role(s) to use only few selected SCCs,
we could remove this patch.