test: make driver operator tests more flexible
Refactor tests so the platform can be set by PlatformStatus structure
instead of just a string for platform type.

This will allow tests to be more flexible when a test needs to use
structures nested in it, for example AzurePlatformStatus which is
platform specific and might affect how CSO decides about starting a
controller.