assets: Remove redundant clusterrole rules
These are already defined in the role and do not need to be duplicated
in the cluster role.

As always, the real changes are in the 'base' directory and the changes
in 'generated' are auto-generated by 'make update'.

While we're here, we also remove some unnecessary quotes for the source
definition.

Signed-off-by: Stephen Finucane <stephenfin@redhat.com>