openstack-cinder: Consume CA cert from credentials secret
cloud-credential-operator now supports deploying the CA cert to the
secrets it generates, which means we can start consuming it from there
rather than from configuration.

While here, we also rename the volume mount uses for the clouds.yaml
file to a more meaningful name and ensure the mounted CA certs (both the
new and legacy ones) are read-only.

As always, the real changes are happening in the 'base', 'hypershift',
and 'standalone' directories of
'assets/csidriveroperators/openstack-cinder'. Everything in 'generated'
is (surprise!) generated by kustomize (using 'make update').

Signed-off-by: Stephen Finucane <stephenfin@redhat.com>