assets: Add control-plane toleration
The 'node-role.kubernetes.io/master' annotation is deprecated in favour
of 'node-role.kubernetes.io/control-plane' [1][2] and currently produces
a warning when modifying resources with oc/kubectl.

  ❯ oc scale -n openshift-cluster-storage-operator deployment cluster-storage-operator --replicas 0
  Warning: spec.template.spec.nodeSelector[node-role.kubernetes.io/master]: use "node-role.kubernetes.io/control-plane" instead
  deployment.apps/cluster-storage-operator scaled

machine-config-operator (MCO) is not currently setting this label, so
we're unable to actually use it for nodeSelector and remove this warning
yet. We can, however, start tolerating it for when MCO starts doing so.

[1] https://github.com/kubernetes/enhancements/tree/master/keps/sig-cluster-lifecycle/kubeadm/2067-rename-master-label-taint
[2] https://kubernetes.io/docs/reference/labels-annotations-taints/#node-role-kubernetes-io-master-taint

Signed-off-by: Stephen Finucane <stephenfin@redhat.com>