Revert "add securityContext.runAsUser=1001 on azure hypershift deployments"
This reverts commit 1faefbcf2e4e7794546d6681f2a4eac13516fa09.

This is needed because runAsUser should now be set dynamically in
HyperShift.

For clusters that have securitycontextconstraint CRD HyperShift will
activate SCC capability for HCP and will not set any default UID for
runAsUser security constraint - allocation is done through SCC as
usual with regular OpenShift deployment.

For clusters without SCC HyperShift will enable UID allocation
for hosted control planes it deploys. The UID to be allocated is
provided as RUN_AS_USER environment variable to CSO which has to
respect this setting and reflect the UID in it's deployment runAsUser
value, and it also has to pass down this value to it's operands
(CSI Operators) which also have to set runAsUser value accordingly.

The reason to revert this commit is not a mistake but a reaction
to this dynamic behavior being introduced in HyperShift.

At the time of merging this revert all storage component patches
are be already merged and should act as described above.