Add VolumeAttributesClass(VAC) alert for parameter mismatch
If there are multiple VACs with different parameters, they can react unexpectedly. We alert the user if there are any mismatches.

Assuming there are multiple VACs on one driver and they have different parameters set. If the user changes from VAC A to VAC B, they would expect to set every non-defined parameter to be set back to the default. That is not the case! The non-defined parameters stay the same. That could lead to confusion later, which is why we need this alert.

assisted by claude