Fix deployment controller Progressing=true on node drain
When a node reboots with drain or any time a CSI driver operator Deployment
misses a replica, CSO should not get Progressing=true.

CSO uses a custom Deployment controller, so fix that one. In the long run,
it would be better to use library-go's DeploymentController, but that would
need much bigger refactoring.